manifests: move the logic for manipulating manifests into a utility class
part of osbs-7170

both group_manifests and floating_tags_push need to be able to manipulate
manifests in similar ways, so move the logic out of group_manifests and
into manifest_util so both can access it cleanly.
